+++ RELEASE_NOTES.md
@@ -21,6 +21,12 @@ _This release remains source compatible for both hosts and managers._
   as keys in associative containers (e.g. `dict`/`std::unordered_map`).
   [#573](https://github.com/OpenAssetIO/OpenAssetIO/issues/573)
 
+- Added bounds checking of array access to convenience methods that wrap
+  batch-first callback-based signatures. This avoids some potential
+  segfaults if a manager plugin calls a callback passing an invalid index
+  argument.
+  [#1196](https://github.com/OpenAssetIO/OpenAssetIO/issues/1196)
+
 - Removed the requirement to define C++ destructors for classes
   inheriting from `LoggerInterface` and
   `ManagerImplementationFactoryInterface`. I.e. the destructors are no